refactor package names try to add scala and java together, also add socket based customized streaming receiver
