+++ spark/spark-streaming-topic/spark-streaming-pro/pom.xml
@@ -13,7 +13,7 @@
         <maven.compiler.target>1.8</maven.compiler.target>
         <spark.version>3.4.1</spark.version>
         <hadoop.client.version>2.7.3</hadoop.client.version>
-        <jackson.version>2.10.3</jackson.version>
+        <jackson.version>2.14.2</jackson.version>
         <gson.version>2.8.0</gson.version>
         <scala.version>2.12.18</scala.version>
         <scala.compat.version>2.12</scala.compat.version>
@@ -62,11 +62,6 @@
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
             <version>${gson.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>${jackson.version}</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
