+++ spark/spark-streaming-topic/spark-streaming-pro/src/test/org/emma/spark/streaming/ContainerTestWrapper.scala
@@ -1,0 +1,9 @@
+package org.emma.spark.streaming
+
+//
+//import org.testcontainers.Network
+
+trait ContainerTestWrapper {
+//  lazy val network: Network
+
+}
