+++ spark/spark-streaming-topic/spark-streaming-pro/src/test/org/emma/spark/streaming/SparkSessionITWrapper.scala
@@ -1,0 +1,15 @@
+package org.emma.spark.streaming
+
+import org.apache.spark.sql.SparkSession
+
+trait SparkSessionITWrapper {
+  lazy val spark: SparkSession = {
+    val spark = SparkSession
+      .builder()
+      .master("local")
+      .appName("Spark Integration Test Wrapper")
+      .getOrCreate()
+    spark.sparkContext.setLogLevel("INFO")
+    spark
+  }
+}
