--- spark/spark-streaming-topic/spark-streaming-pro/src/test/scala/org/emma/spark/streaming/topic/AggErrorCountIT.scala
@@ -1,12 +1,0 @@
-//package org.emma.spark.streaming.topic
-//
-//
-//import org.apache.spark.sql.SparkSession
-//import org.sca
-//import org.scalatest
-//import org.scalatest.{BeforeAndAfterAll, FlatSpec}
-//class AggErrorCountIT extends FlatSpec with ForAllTestContainer with BeforeAndAfterAll with ContainerTestWrapper with SparkSessionITWrapper {
-//  val sparkTest: SparkSession = spark
-//
-//
-//}
