--- spark/spark-streaming-topic/spark-streaming-pro/src/test/scala/org/emma/spark/streaming/topic/ContainerTestWrapper.scala
@@ -1,8 +1,0 @@
-package org.emma.spark.streaming.topic
-//
-//import org.testcontainers.Network
-
-trait ContainerTestWrapper {
-//  lazy val network: Network
-
-}
